backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pymongo>=4.0
transformers>=4.30
torch>=2.0
python-dotenv>=1.0
python-multipart>=0.0.6
starlette==0.27.0
librosa>=0.10
soundfile>=0.12
whisper
